+++ app/domain/entities/travel_entity.go
@@ -1,6 +1,6 @@
-package main
+package entity
 
-type TravelModel struct {
+type TravelEntity struct {
 	Country string `json:"country"`
 	City    string `json:"city"`
 }
