+++ app/main.go
@@ -1,4 +1,6 @@
 package main
+package GenerateTravel
+
 
 import (
 	"context"
@@ -12,21 +14,10 @@ import (
 	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
-type TravelModel struct {
-	Country string `json:"country"`
-	City    string `json:"city"`
-}
 
-func generateTravel() TravelModel {
-	var travel = TravelModel{
-		Country: "Portugal",
-		City:    "Lisbon",
-	}
-	return travel
-}
 
 func fetchTravel(c *gin.Context) {
-	c.IndentedJSON(http.StatusOK, generateTravel())
+	c.IndentedJSON(http.StatusOK, GenerateTravel())
 }
 
 func runApp(data any) {
@@ -36,22 +27,6 @@ func runApp(data any) {
 }
 
 func main() {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	client, e := mongo.Connect(ctx, options.Client().ApplyURI("mongodb://localhost:27017"))
-	if e != nil {
-		client.Disconnect(ctx)
-		panic(e)
-	}
-	collection := client.Database("travels").Collection("travel")
-	result, err := collection.InsertOne(context.Background(), generateTravel())
-	if err != nil {
-		panic(err)
-	}
-	dataFromDb, _e := collection.Find(ctx, bson.M{})
-	if _e != nil {
-		panic(e)
-	}
-	fmt.Print(result.InsertedID)
+	
 	runApp(dataFromDb)
 }
